cancel button 추가, check 이미지뷰 추가 및 글자 크기 조정
