+++ Projects/App/Sources/Application/AppDelegate.swift
@@ -1,10 +1,3 @@
-//
-//  AppDelegate.swift
-//  ProjectDescriptionHelpers
-//
-//  Created by 강민성 on 10/30/23.
-//
-
 import UIKit
 
 @main
