+++ Projects/Core/Sources/CoreExport.swift
@@ -1,9 +1,2 @@
-//
-//  CoreExport.swift
-//  ProjectDescriptionHelpers
-//
-//  Created by 강민성 on 10/30/23.
-//
-
 @_exported import CoreNetwork
 @_exported import CoreNetworkInterface
