+++ Projects/Shared/DSKit/Sources/source.swift
@@ -1,1 +1,2 @@
-// this is for tuist
+@_exported import SharedUtils
+@_exported import SharedUtilsInterface
